requirements-dev.txt
aiosqlite==0.17.0
alembic==1.7.5
anyio==3.5.0
asgiref==3.5.0
asyncio==3.4.3
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coverage==6.3
cryptography==36.0.1
databases==0.5.4
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
fastapi-users==9.2.2
fastapi-users-db-ormar==1.0.0
ghp-import==2.0.2
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
importlib-metadata==4.10.1
iniconfig==1.1.1
Jinja2==3.0.3
makefun==1.13.1
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.3
mkdocs-material==8.1.8
mkdocs-material-extensions==1.0.3
ormar==0.10.24
packaging==21.3
parameterized==0.8.1
passlib==1.7.4
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
pymdown-extensions==9.1
pyparsing==3.0.7
pytest==6.2.5
pytest-asyncio==0.17.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0.post1
uvloop==0.16.0
watchdog==2.1.6
watchgod==0.7
websockets==10.1
zipp==3.7.0
